+++ packages/core/src/lib/modules/wallet-instance.ts
